alteracoes editar imagem favorita

+++ src/app/modules/produto/editar-produto/editar-imagens-produto/editar-imagens-produto.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { DomSanitizer, SafeResourceUrl } from '@angular/platform-browser';
 import { ActivatedRoute, Router } from '@angular/router';
-import {  ToastrService } from 'ngx-toastr';
+import { ToastrService } from 'ngx-toastr';
 import { ProdutoService } from 'src/app/services/produto.service';
 import { Imagem, Produto } from '../../models/Produto';
 
@@ -23,38 +23,38 @@ export class EditarImagensProdutoComponent implements OnInit {
     this.route.params.subscribe(parametros => {
       this.id = parametros['id'];
     });
+
+  }
+
+  ngOnInit() {
     this.produtoService.getProdutoById(this.id).subscribe(resp => {
       this.produto = resp;
     })
     this.produtoService.getImagensProduto(this.id).subscribe(response => {
       this.imagens = response;
-      response.forEach(element =>{
-        const t=this.sanitizer.bypassSecurityTrustResourceUrl(`data:image/png;base64, ${element.imagem}`)
-        let img=new Imagem();
-        img.id=element.id;
-        img.imagem=element.imagem;
-        img.imageToShow=t;
-        img.caminho=element.caminho;
-        img.imagemPrincipal=element.imagemPrincipal;
+      response.forEach(element => {
+        const t = this.sanitizer.bypassSecurityTrustResourceUrl(`data:image/png;base64, ${element.imagem}`)
+        let img = new Imagem();
+        img.id = element.id;
+        img.imagem = element.imagem;
+        img.imageToShow = t;
+        img.caminho = element.caminho;
+        img.imagemPrincipal = element.imagemPrincipal;
         this.imageToShow.push(img);
       }
       )
     })
-    console.log(this.imageToShow);
-  }
-
-  ngOnInit() {
   }
 
   enviar() {
     for (const file of this.files) {
       this.imagensUpload.append('file', file);
     }
-    this.produtoService.postFotoProduto(this.imagensUpload, this.produto.id!, 1).subscribe(response => {
-      this.toastr.success("Imagens adicionadas com sucesso", "Ok",{
+    this.produtoService.postFotoProduto(this.imagensUpload, this.produto.id!, -1).subscribe(response => {
+      this.toastr.success("Imagens adicionadas com sucesso", "Ok", {
         timeOut: 3000, positionClass: 'toast-top-center',
-        });
-        window.location.reload();
+      });
+      window.location.reload();
     })
     this.novasImagems.length = null;
   }
@@ -80,13 +80,38 @@ export class EditarImagensProdutoComponent implements OnInit {
     this.files = this.fileList;
     this.fileList = null;
   }
-  deleteImageBanco(i: number): void {
-    let img = this.imagens[i];
-      this.produtoService.deleteImagensProduto(img.id!).subscribe(response => {
-        window.location.reload();
+  deleteImageBanco(img: Imagem): void {
+    if (img.imagemPrincipal) {
+      this.toastr.warning("Não é possível deletar uma imagem favoritada", "Atenção", {
+        timeOut: 3000, positionClass: 'toast-top-center',
       });
+      return;
+    }
+    this.produtoService.deleteImagensProduto(img.id!).subscribe(response => {
+      window.location.reload();
+    });
   }
+
   backProdutos() {
     this.router.navigate(['produtos/editar', this.produto.id]);
+  }
+  public toggleSelected(img: Imagem) {
+    for (let imgP of this.imageToShow) {
+      if (imgP.imagemPrincipal) {
+        imgP.imagemPrincipal = false;
+      }
+    }
+    img.imagemPrincipal = true;
+    this.produtoService.postFotoFavorita(img.id!, this.produto.id!).subscribe(response => {
+      this.toastr.success("Nova imagem favoritada", "Ok", {
+        timeOut: 3000, positionClass: 'toast-top-center',
+      })},
+       err=>{
+      this.toastr.error("Erro ao Favoritar Imagem", "Error", {
+        timeOut: 3000, positionClass: 'toast-top-center',
+      }
+      )}
+
+    );
   }
 }
